requirements.txt
sentence-transformers>=2.2.2
redis>=4.5.0
PyPDF2>=3.0.0
pdfplumber>=0.9.0
mcp>=0.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
PyYAML>=6.0
click>=8.0.0
tqdm>=4.64.0
numpy>=1.24.0
nltk>=3.8.0